fix: headerをcomponentへ切り出し
