+++ src/layouts/default.vue
@@ -1,54 +1,7 @@
 <template>
   <div>
-    <nav
-      class="navbar header has-shadow is-primary"
-      role="navigation"
-      aria-label="main navigation"
-    >
-      <div class="navbar-brand">
-        <a class="navbar-item is-primary" href="/">
-          <span class="is-size-4 has-text-weight-bold">Slide-Share-Spot</span>
-        </a>
-        <div class="navbar-burger">
-          <span />
-          <span />
-          <span />
-        </div>
-      </div>
-      <div class="navbar-end">
-        <div class="navbar-item">
-          <div v-if="!isLogin" class="buttons">
-            <nuxt-link to="/auth/signup" class="button is-primary">
-              <strong>Sign up</strong>
-            </nuxt-link>
-            <nuxt-link to="/auth/login" class="button is-light">
-              Log in
-            </nuxt-link>
-          </div>
-          <div v-else class="buttons">
-            <nuxt-link to="/post" class="button is-light">Post</nuxt-link>
-            <button class="button is-light" @click="logout()">Log out</button>
-          </div>
-        </div>
-      </div>
-    </nav>
-
+    <VHeader />
     <section class="main-content columns">
-      <!-- 
-      <aside class="column is-2 section">
-        <p>username: {{ username }}</p>
-        <ul class="menu-list">
-          <li v-for="(item, key) of items" :key="key">
-            <nuxt-link :to="item.to" exact-active-class="is-active">
-              <b-icon :icon="item.icon" />
-              {{ item.title }}
-            </nuxt-link>
-          </li>
-        </ul>
-      </aside>
-
-      -->
-
       <div class="container column is-10">
         <nuxt />
       </div>
@@ -57,7 +10,11 @@
 </template>
 
 <script>
+import VHeader from '~/components/VHeader'
 export default {
+  components: {
+    VHeader
+  },
   data() {
     return {
       items: [
@@ -77,24 +34,6 @@ export default {
           to: { name: 'article-search' }
         }
       ]
-    }
-  },
-  computed: {
-    username() {
-      return this.$store.getters.username
-    },
-    isLogin() {
-      return this.$store.getters.isAuthenticated
-    }
-  },
-  methods: {
-    async logout() {
-      await this.$store.dispatch('logout')
-      this.$buefy.toast.open({
-        message: 'ログoutできました',
-        type: 'is-success'
-      })
-      this.$router.push('/')
     }
   }
 }
